refactor(bridge): extract forwarding messages
